+++ lib/enemy/extensions.dart
@@ -45,7 +45,7 @@ extension EnemyExtensions on Enemy {
   /// Add in the game a text with animation representing damage received
   void showDamage(
     double damage, {
-    TextConfig? config,
+    TextPaintConfig? config,
     double initVelocityTop = -5,
     double gravity = 0.5,
     double maxDownSize = 20,
@@ -60,7 +60,7 @@ extension EnemyExtensions on Enemy {
           position.top,
         ),
         config: config ??
-            TextConfig(
+            TextPaintConfig(
               fontSize: 14,
               color: Colors.white,
             ),