+++ lib/util/mixins/sensor.dart
@@ -1,6 +1,7 @@
 import 'dart:ui';
 
 import 'package:bonfire/base/game_component.dart';
+import 'package:bonfire/bonfire.dart';
 import 'package:bonfire/collision/collision_area.dart';
 import 'package:bonfire/collision/collision_config.dart';
 import 'package:bonfire/collision/object_collision.dart';
@@ -12,11 +13,13 @@ import '../interval_tick.dart';
 Paint _paintSensor = Paint()..color = Colors.red.withOpacity(0.5);
 
 mixin Sensor on GameComponent {
-  void onContact(GameComponent collision);
+  void onContact(GameComponent component);
 
   int _intervalCheckContact = 0;
   IntervalTick? _tick;
   Vector2Rect? _sensorArea;
+
+  CollisionConfig? _collisionConfig;
 
   Vector2Rect get sensorArea {
     if (_sensorArea != null) {
@@ -33,10 +36,24 @@ mixin Sensor on GameComponent {
   void setupSensorArea(Vector2Rect s, {int intervalCheck = 250}) {
     _sensorArea = s;
     _intervalCheckContact = intervalCheck;
+    _collisionConfig = CollisionConfig(
+      collisions: [
+        CollisionArea.fromVector2Rect(rect: s),
+      ],
+    );
   }
 
   @override
   void update(double dt) {
+    if (_collisionConfig == null) {
+      _collisionConfig = CollisionConfig(
+        collisions: [
+          CollisionArea.fromVector2Rect(
+            rect: Vector2Rect(Vector2.zero(), position.size),
+          ),
+        ],
+      );
+    }
     if (_tick == null || _tick?.interval != _intervalCheckContact) {
       _tick = IntervalTick(_intervalCheckContact, tick: _verifyContact);
     } else {
@@ -54,12 +71,13 @@ mixin Sensor on GameComponent {
   }
 
   void _verifyContact() {
+    _collisionConfig?.updatePosition(sensorArea);
     for (final i in gameRef.visibleComponents()) {
-      if (i is ObjectCollision) {
-        CollisionConfig config = CollisionConfig(
-          collisions: [CollisionArea.fromVector2Rect(rect: sensorArea)],
-        )..updatePosition(sensorArea);
-        if (i.collisionConfig?.verifyCollision(config) ?? false) {
+      if (i.isObjectCollision()) {
+        if ((i as ObjectCollision)
+                .collisionConfig
+                ?.verifyCollision(_collisionConfig) ??
+            false) {
           onContact(i);
         }
       } else if (i.position.overlaps(sensorArea)) {